+++ pkg/services.go
@@ -3,6 +3,7 @@ package exporter
 import (
 	"context"
 	"fmt"
+	"regexp"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -10,6 +11,8 @@ import (
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/databasemigrationservice"
 	"github.com/aws/aws-sdk-go/service/ec2"
+	"github.com/aws/aws-sdk-go/service/prometheusservice"
+	"github.com/aws/aws-sdk-go/service/storagegateway"
 )
 
 type ResourceFunc func(context.Context, tagsInterface, *Job, string) ([]*taggedResource, error)
@@ -19,7 +22,6 @@ type FilterFunc func(context.Context, tagsInterface, []*taggedResource) ([]*tagg
 type serviceFilter struct {
 	Namespace        string
 	Alias            string
-	IgnoreLength     bool
 	ResourceFilters  []*string
 	DimensionRegexps []*string
 	ResourceFunc     ResourceFunc
@@ -106,7 +108,8 @@ var (
 				})
 				for _, resource := range inputResources {
 					for i, gw := range output.Items {
-						if strings.Contains(resource.ARN, *gw.Id) {
+						searchString := regexp.MustCompile(fmt.Sprintf(".*apis/%s$", *gw.Id))
+						if searchString.MatchString(resource.ARN) {
 							r := resource
 							r.ARN = strings.ReplaceAll(resource.ARN, *gw.Id, *gw.Name)
 							outputResources = append(outputResources, r)
@@ -122,6 +125,9 @@ var (
 			Alias:     "mq",
 			ResourceFilters: []*string{
 				aws.String("mq"),
+			},
+			DimensionRegexps: []*string{
+				aws.String("broker:(?P<Broker>[^:]+)"),
 			},
 		}, {
 			Namespace: "AWS/AppSync",
@@ -179,9 +185,8 @@ var (
 			Alias:     "beanstalk",
 		},
 		{
-			Namespace:    "AWS/Billing",
-			Alias:        "billing",
-			IgnoreLength: true,
+			Namespace: "AWS/Billing",
+			Alias:     "billing",
 		}, {
 			Namespace: "AWS/Cassandra",
 			Alias:     "cassandra",
@@ -291,7 +296,7 @@ var (
 				aws.String("directconnect"),
 			},
 			DimensionRegexps: []*string{
-				aws.String("dxcon:(?P<ConnectionId>[^/]+)"),
+				aws.String(":dxcon/(?P<ConnectionId>[^/]+)"),
 			},
 		}, {
 			Namespace: "AWS/DynamoDB",
@@ -411,6 +416,15 @@ var (
 				aws.String("cluster/(?P<JobFlowId>[^/]+)"),
 			},
 		}, {
+			Namespace: "AWS/EMRServerless",
+			Alias:     "emr-serverless",
+			ResourceFilters: []*string{
+				aws.String("emr-serverless:applications"),
+			},
+			DimensionRegexps: []*string{
+				aws.String("applications/(?P<ApplicationId>[^/]+)"),
+			},
+		}, {
 			Namespace: "AWS/ES",
 			Alias:     "es",
 			ResourceFilters: []*string{
@@ -489,6 +503,15 @@ var (
 				aws.String(":cluster/(?P<Cluster_Name>[^/]+)"),
 			},
 		}, {
+			Namespace: "AWS/KafkaConnect",
+			Alias:     "kafkaconnect",
+			ResourceFilters: []*string{
+				aws.String("kafkaconnect"),
+			},
+			DimensionRegexps: []*string{
+				aws.String(":connector/(?P<Connector_Name>[^/]+)"),
+			},
+		}, {
 			Namespace: "AWS/Kinesis",
 			Alias:     "kinesis",
 			ResourceFilters: []*string{
@@ -508,6 +531,15 @@ var (
 			},
 			DimensionRegexps: []*string{
 				aws.String(":function:(?P<FunctionName>[^/]+)"),
+			},
+		}, {
+			Namespace: "AWS/MediaTailor",
+			Alias:     "mediatailor",
+			ResourceFilters: []*string{
+				aws.String("mediatailor:playbackConfiguration"),
+			},
+			DimensionRegexps: []*string{
+				aws.String("playbackConfiguration/(?P<ConfigurationName>[^/]+)"),
 			},
 		}, {
 			Namespace: "AWS/Neptune",
@@ -568,6 +600,35 @@ var (
 				aws.String(":vpc-endpoint-service:(?P<Service_Id>.+)"),
 			},
 		}, {
+			Namespace: "AWS/Prometheus",
+			Alias:     "amp",
+			ResourceFunc: func(ctx context.Context, iface tagsInterface, job *Job, region string) (resources []*taggedResource, err error) {
+				pageNum := 0
+				return resources, iface.prometheusClient.ListWorkspacesPagesWithContext(ctx, &prometheusservice.ListWorkspacesInput{},
+					func(page *prometheusservice.ListWorkspacesOutput, more bool) bool {
+						pageNum++
+						managedPrometheusAPICounter.Inc()
+
+						for _, ws := range page.Workspaces {
+							resource := taggedResource{
+								ARN:       aws.StringValue(ws.Arn),
+								Namespace: job.Type,
+								Region:    region,
+							}
+
+							for key, value := range ws.Tags {
+								resource.Tags = append(resource.Tags, Tag{Key: key, Value: *value})
+							}
+
+							if resource.filterThroughTags(job.SearchTags) {
+								resources = append(resources, &resource)
+							}
+						}
+						return pageNum < 100
+					},
+				)
+			},
+		}, {
 			Namespace: "AWS/RDS",
 			Alias:     "rds",
 			ResourceFilters: []*string{
@@ -606,9 +667,8 @@ var (
 				aws.String(":healthcheck/(?P<HealthCheckId>[^/]+)"),
 			},
 		}, {
-			Namespace:    "AWS/S3",
-			Alias:        "s3",
-			IgnoreLength: true,
+			Namespace: "AWS/S3",
+			Alias:     "s3",
 			ResourceFilters: []*string{
 				aws.String("s3"),
 			},
@@ -644,6 +704,50 @@ var (
 			},
 			DimensionRegexps: []*string{
 				aws.String("(?P<QueueName>[^:]+)$"),
+			},
+		}, {
+			Namespace: "AWS/StorageGateway",
+			Alias:     "storagegateway",
+			ResourceFilters: []*string{
+				aws.String("storagegateway"),
+			},
+			DimensionRegexps: []*string{
+				aws.String(":gateway/(?P<GatewayId>[^:]+)$"),
+				aws.String(":share/(?P<ShareId>[^:]+)$"),
+				aws.String("^(?P<GatewayId>[^:/]+)/(?P<GatewayName>[^:]+)$"),
+			},
+			ResourceFunc: func(ctx context.Context, iface tagsInterface, job *Job, region string) (resources []*taggedResource, err error) {
+				pageNum := 0
+				return resources, iface.storagegatewayClient.ListGatewaysPagesWithContext(ctx, &storagegateway.ListGatewaysInput{},
+					func(page *storagegateway.ListGatewaysOutput, more bool) bool {
+						pageNum++
+						storagegatewayAPICounter.Inc()
+
+						for _, gwa := range page.Gateways {
+							resource := taggedResource{
+								ARN:       fmt.Sprintf("%s/%s", *gwa.GatewayId, *gwa.GatewayName),
+								Namespace: job.Type,
+								Region:    region,
+							}
+
+							tagsRequest := &storagegateway.ListTagsForResourceInput{
+								ResourceARN: gwa.GatewayARN,
+							}
+							tagsResponse, _ := iface.storagegatewayClient.ListTagsForResource(tagsRequest)
+							storagegatewayAPICounter.Inc()
+
+							for _, t := range tagsResponse.Tags {
+								resource.Tags = append(resource.Tags, Tag{Key: *t.Key, Value: *t.Value})
+							}
+
+							if resource.filterThroughTags(job.SearchTags) {
+								resources = append(resources, &resource)
+							}
+						}
+
+						return pageNum < 100
+					},
+				)
 			},
 		}, {
 			Namespace: "AWS/TransitGateway",