+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+# 0.42.0-alpha
+
+* Resolve logging issue (@datsabk)
+* MediaTailor - Correct dimension regex for MT (@scott-mccracken)
+* Helm chart update for optional test-connection pod (@nikosmeds)
+* Helm chart update to set priorityClassName (@nikosmeds)
+* Bump github.com/aws/aws-sdk-go from 1.44.122 to 1.44.127
+* Bump github.com/urfave/cli/v2 from 2.20.3 to 2.23.0
+
+# 0.41.0-alpha
+
+* Clean up unused variables. (@cristiangreco)
+* Fix typo: sts-endpoint should be sts-region. (@cristiangreco)
+* Enabled Managed prometheus metrics (@datsabk)
+* Add support for AWS Kafka Connect (@cgowthaman)
+* Import CloudWatch mixin. (@jeschkies)
+* main.go refactoring: define cmd action as a separate func. (@cristiangreco)
+* Add support for EMR Serverless (@cgowthaman)
+
+# 0.40.0-alpha
+* Fix typo in Charts.yml (@yasharne)
+* Subcommand `verify-config` actually validates the config file. (@cristiangreco)
+* Add dimensions regex for AmazonMQ. (@cristiangreco)
+* Fix metrics with additional dimensions being not being scraped. (@cristiangreco)
+* Remove unused code, add test for RemoveDuplicateMetrics. (@cristiangreco)
+* Bump github.com/sirupsen/logrus
+* Bump github.com/urfave/cli/v2
+* Bump github.com/aws/aws-sdk-go
+* Bump actions/setup-python
+
+# 0.39.0-alpha
+* Improve code quality and unblock this release (cristiangreco)
+* Add helm chart (vkobets)
+* Fix DX metrics (paulojmdias)
+* Fix searchTags and bad dimension name (femiagbabiaka)
+* Handle empty list in filter metric tests (mtt88)
+* Add AWS Elemental MediaTailor support (scott-mccracken)
+* Support storagegateway metrics (sedan07)
+* Filter api gateway resources to skip "stages" (ch4rms)
+* Bump aws-sdk, urfave/cli, prometheus/client_golang
+
 # 0.38.0-alpha
 
 * Set max page size for tagging API requests (#617)